refactor: move triggers to pipeline resource
