+++ opentoolchain/resource_opentoolchain_tekton_pipeline.go
@@ -7,6 +7,7 @@ import (
 	uuid "github.com/google/uuid"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 	"strings"
 )
 
@@ -84,6 +85,79 @@ func resourceOpenToolchainTektonPipeline() *schema.Resource {
 					},
 				},
 			},
+			"trigger": {
+				Type:     schema.TypeSet,
+				Required: true,
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"id": {
+							Description: "Trigger ID",
+							Type:        schema.TypeString,
+							Computed:    true,
+						},
+						"enabled": {
+							Description: "`true` if trigger should be active",
+							Type:        schema.TypeBool,
+							Optional:    true,
+							Default:     true,
+						},
+						"github_integration_id": {
+							Description: "Github integration ID",
+							Type:        schema.TypeString,
+							Optional:    true,
+						},
+						"github_url": {
+							Description: "Github repository URL",
+							Type:        schema.TypeString,
+							Optional:    true,
+						},
+						"name": {
+							Description: "Trigger name",
+							Type:        schema.TypeString,
+							Required:    true,
+						},
+						"event_listener": {
+							Description: "Event Listener name (from .tekton pipeline definition)",
+							Type:        schema.TypeString,
+							Required:    true,
+						},
+						"on_pull_request": {
+							Description: "Trigger when pull request is opened or updated",
+							Type:        schema.TypeBool,
+							Optional:    true,
+							Default:     false,
+						},
+						"on_pull_request_closed": {
+							Description: "Trigger when pull request is closed",
+							Type:        schema.TypeBool,
+							Optional:    true,
+							Default:     false,
+						},
+						"on_push": {
+							Description: "Trigger when commit is pushed",
+							Type:        schema.TypeBool,
+							Optional:    true,
+							Default:     false,
+						},
+						"branch": {
+							Description: "GitHub branch",
+							Type:        schema.TypeString,
+							Optional:    true,
+						},
+						"pattern": {
+							Description: "GitHub branch pattern, if `branch` is not specified, otherwise setting is ignored",
+							Type:        schema.TypeString,
+							Optional:    true,
+						},
+						"type": {
+							Description:  "Trigger type",
+							Type:         schema.TypeString,
+							ValidateFunc: validation.StringInSlice([]string{"scm", "manual"}, false),
+							Required:     true,
+						},
+					},
+				},
+			},
 		},
 	}
 }
@@ -93,6 +167,7 @@ func resourceOpenToolchainTektonPipelineCreate(ctx context.Context, d *schema.Re
 	toolchainID := d.Get("toolchain_id").(string)
 	name := d.Get("name").(string)
 	inputs := d.Get("definition").(*schema.Set)
+	triggers := d.Get("trigger").(*schema.Set)
 
 	config := m.(*ProviderConfig)
 	c := config.OTClient
@@ -184,6 +259,7 @@ func resourceOpenToolchainTektonPipelineCreate(ctx context.Context, d *schema.Re
 		EnvID:                &envID,
 		PipelineDefinitionID: definition.Definition.ID,
 		Inputs:               definition.Inputs,
+		Triggers:             expandTektonPipelineTriggers(triggers.List()),
 		Worker: &oc.PatchTektonPipelineParamsWorker{
 			// current pipeline defaults
 			WorkerID:   getStringPtr("public"),
@@ -243,6 +319,10 @@ func resourceOpenToolchainTektonPipelineRead(ctx context.Context, d *schema.Reso
 	}
 
 	if err = d.Set("definition", flattenTektonPipelineDefinition(pipeline.Inputs)); err != nil {
+		return diag.Errorf("Error setting pipeline definition inputs: %s", err)
+	}
+
+	if err = d.Set("trigger", flattenTektonPipelineTriggers(pipeline.Triggers)); err != nil {
 		return diag.Errorf("Error setting pipeline definition inputs: %s", err)
 	}
 
@@ -327,8 +407,13 @@ func resourceOpenToolchainTektonPipelineUpdate(ctx context.Context, d *schema.Re
 		patchOptions.Inputs = definition.Inputs
 	}
 
+	if d.HasChange("trigger") {
+		triggers := d.Get("trigger").(*schema.Set)
+		patchOptions.Triggers = expandTektonPipelineTriggers(triggers.List())
+	}
+
 	// add other conditions here
-	if d.HasChange("definition") {
+	if d.HasChange("definition") || d.HasChange("trigger") {
 		_, _, err := c.PatchTektonPipelineWithContext(ctx, patchOptions)
 
 		if err != nil {
@@ -365,6 +450,53 @@ func expandTektonPipelineDefinitionInputs(c *oc.OpenToolchainV1, envID *string, 
 	return result
 }
 
+func expandTektonPipelineTriggers(t []interface{}) []oc.TektonPipelineTrigger {
+	result := make([]oc.TektonPipelineTrigger, len(t))
+
+	for index, trig := range t {
+		trigger := trig.(map[string]interface{})
+		name := trigger["name"].(string)
+		eventListener := trigger["event_listener"].(string)
+		triggerType := trigger["type"].(string)
+		enabled := trigger["enabled"].(bool)
+
+		result[index] = oc.TektonPipelineTrigger{
+			ID:            getStringPtr(uuid.NewString()),
+			Name:          &name,
+			EventListener: &eventListener,
+			Type:          &triggerType,
+			Disabled:      getBoolPtr(!enabled),
+		}
+
+		if triggerType == "scm" {
+			githubIntegrationID := trigger["github_integration_id"].(string)
+			onPush := trigger["on_push"].(bool)
+			onPR := trigger["on_pull_request"].(bool)
+			onPRClosed := trigger["on_pull_request_closed"].(bool)
+			branch := trigger["branch"].(string)
+			pattern := trigger["pattern"].(string)
+			url := trigger["github_url"].(string)
+
+			result[index].ServiceInstanceID = &githubIntegrationID
+
+			result[index].ScmSource = &oc.TektonPipelineTriggerScmSource{
+				URL:     &url,
+				Type:    getStringPtr("GitHub"), // TODO: what other types available?
+				Branch:  &branch,
+				Pattern: &pattern,
+			}
+
+			result[index].Events = &oc.TektonPipelineTriggerEvents{
+				Push:              &onPush,
+				PullRequest:       &onPR,
+				PullRequestClosed: &onPRClosed,
+			}
+		}
+	}
+
+	return result
+}
+
 func flattenTektonPipelineDefinition(d []oc.TektonPipelineInput) []interface{} {
 	var result []interface{}
 
@@ -379,6 +511,34 @@ func flattenTektonPipelineDefinition(d []oc.TektonPipelineInput) []interface{} {
 
 			result = append(result, input)
 		}
+	}
+
+	return result
+}
+
+func flattenTektonPipelineTriggers(t []oc.TektonPipelineTrigger) []interface{} {
+	var result []interface{}
+
+	for _, trg := range t {
+		trigger := map[string]interface{}{
+			"id":             *trg.ID,
+			"enabled":        !*trg.Disabled,
+			"name":           *trg.Name,
+			"event_listener": *trg.EventListener,
+			"type":           *trg.Type,
+		}
+
+		if *trg.Type == "scm" {
+			trigger["github_integration_id"] = *trg.ServiceInstanceID
+			trigger["github_url"] = *trg.ScmSource.URL
+			trigger["on_pull_request"] = *trg.Events.PullRequest
+			trigger["on_pull_request_closed"] = *trg.Events.PullRequestClosed
+			trigger["on_push"] = *trg.Events.Push
+			trigger["branch"] = *trg.ScmSource.Branch
+			trigger["pattern"] = *trg.ScmSource.Pattern
+		}
+
+		result = append(result, trigger)
 	}
 
 	return result
